Rename ConfigReader::Wicked to ConfigReader::Sysconfig
